+++ libraries/mongodb_config_helpers.rb
@@ -16,4 +16,23 @@ module MongoDBConfigHelpers
     end \
           .compact.join("\n")
   end
+
+  def to_yaml_options(config)
+    config.to_hash.compact.to_yaml
+  end
+end
+
+# Monkey patches Hash to allow us to throw away keys that have empty or nil values
+class Hash
+  def compact
+    inject({}) do |new_hash, (k, v)|
+      if v.is_a?(Hash)
+        v = v.compact
+        new_hash[k] = v unless v.empty?
+      else
+        new_hash[k] = v unless v.nil?
+      end
+      new_hash
+    end
+  end
 end