+++ recipes/shard.rb
@@ -6,6 +6,8 @@
 # Authors:
 #       Markus Korn <markus.korn@edelight.de>
 #
+# Copyright 2016-2017, Grant Ridder
+#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -19,22 +21,23 @@
 # limitations under the License.
 #
 
-node.set['mongodb']['is_shard'] = true
-node.set['mongodb']['shard_name'] = node['mongodb']['shard_name']
-node.set['mongodb']['is_replicaset'] = node['mongodb']['is_replicaset']
-node.set['mongodb']['cluster_name'] = node['mongodb']['cluster_name']
+node.normal['mongodb']['is_shard'] = true
+node.normal['mongodb']['shard_name'] = node['mongodb']['shard_name']
+node.normal['mongodb']['is_replicaset'] = node['mongodb']['is_replicaset']
+node.normal['mongodb']['cluster_name'] = node['mongodb']['cluster_name']
 
 include_recipe 'sc-mongodb::install'
 
-# we are not starting the shard service with the --shardsvr
-# commandline option because right now this only changes the port it's
-# running on, and we are overwriting this port anyway.
-mongodb_instance node['mongodb']['instance_name'] do
-  mongodb_type 'shard'
-  port         node['mongodb']['config']['port']
-  logpath      node['mongodb']['config']['logpath']
-  dbpath       node['mongodb']['config']['dbpath']
-  replicaset   node if node['mongodb']['is_replicaset']
-  enable_rest  node['mongodb']['config']['rest']
-  smallfiles   node['mongodb']['config']['smallfiles']
+ruby_block 'chef_gem_at_converge_time' do
+  block do
+    node['mongodb']['ruby_gems'].each do |gem, version|
+      version = Gem::Dependency.new(gem, version)
+      Chef::Provider::Package::Rubygems::GemEnvironment.new.install(version)
+    end
+  end
+end
+
+mongodb_instance node['mongodb']['instance_name']['mongod'] do
+  mongodb_type 'mongod'
+  replicaset true
 end