+++ test/integration/configserver/inspec/default_spec.rb
@@ -1,6 +1,4 @@
-# mongodb | mongod
-
-describe service('mongodb') do
+describe service('mongod') do
   it { should be_installed }
   it { should be_enabled }
   it { should be_running }