+++ test/integration/default/inspec/default_spec.rb
@@ -1,4 +1,4 @@
-describe service('mongodb') do
+describe service('mongod') do
   it { should be_installed }
   it { should be_enabled }
   it { should be_running }